Add seamless looping and transition between clips (#133)
This PR adds two creation parameters (seamless and mimeCodec), and two actions (set_loops and set_next_file) to the video component. These are meant to be used for playing a looped background during a spoken part of a song, and then transitioning to the next part the next time the loop ends (vamping).

seamless uses the MediaSource API to provide a glitchless transition between clips and loops. The tradeoff is that it is very particular with what formats are supported, and you have to provide the encoding information yourself via mimeCodec.

set_loops will set the remaining number of loops, including the current one. Most often used to turn on (0) or off (1) looping.

set_next_file will set the file to be played after the current one. Multiple calls will replace the file rather than queue up multiple.

An example of the feature in an opus:

  "1":
    prompt: "Start looped video (mp4)"
    next: "2"
    actions:
      - target: "video"
        cmd: "create"
        assets:
          - path: "video_start.mp4"
        params:
          entityId: "mp4_video"
          layer: 2
          visible: true
          looping: 0
          seamless: true
          mimeCodec: 'video/mp4; codecs="avc1.640029, mp4a.40.2"'
      - target: "video"
        cmd: "set_next_file"
        assets:
          - path: "video_looping_middle.mp4"
        params:
          entityId: "mp4_video"
  "2":
    prompt: "End looped video (mp4)"
    next: "3"
    actions:
      - target: "video"
        cmd: "set_next_file"
        assets:
          - path: "video_end.mp4"
        params:
          entityId: "mp4_video"
      - target: "video"
        cmd: "set_loops"
        params:
          entityId: "mp4_video"
          looping: 1

For audio, this PR adds the actions and a seamless parameter only. Here we use the web audio API instead, and we should in theory be able to always use it without drawbacks. The reasons for keeping the old way around are:

    There may be some bugs in the implementation that we discover too late, so it's good to have a safety hatch
    I left some features unimplemented, to be able to focus on other things.